git-svn-id: svn://svn.iws.uni-stuttgart.de/DUMUX/dumux/releases/2.0@5213 2fb0f335-1f38-0410-981e-8018bf24f1b0

+++ appl/lecture/msm/buckleyleverett/buckleyleverett_ff.cc
@@ -77,8 +77,8 @@ int main(int argc, char** argv)
         Dune::FieldVector<Scalar, dim> upperRight(300);
         upperRight[1] = 75;
 
-        int cellNumberX = 300/discretizationLength;
-        int cellNumberY = 75/discretizationLength;
+        int cellNumberX = static_cast<int>(300/discretizationLength);
+        int cellNumberY = static_cast<int>(75/discretizationLength);
 
         Dune::FieldVector<int, dim> cellNumbers(cellNumberX);
         cellNumbers[1] = cellNumberY;
